+++ lib/src/event_type.dart
@@ -27,6 +27,9 @@ class EventType {
   /// Event sent when deleting a reaction
   static const String reactionDeleted = 'reaction.deleted';
 
+  /// Event sent when updating a reaction
+  static const String reactionUpdated = 'reaction.updated';
+
   /// Event sent when updating a message
   static const String messageUpdated = 'message.updated';
 
@@ -57,6 +60,9 @@ class EventType {
 
   /// Event sent when a channel is updated
   static const String channelUpdated = 'channel.updated';
+
+  /// Event sent when a user is updated
+  static const String userUpdated = 'user.updated';
 
   /// Event sent when a member is added to a channel
   static const String memberAdded = 'member.added';